Better organize asdf extension
