+++ specutils/io/asdf/spectra.py
